move test cases.
